requirements.txt
fastapi[standard]
pydantic
uvicorn
requests
python-jose
auth0-python
pymongo